fix!: corrige entidade billing
BREAKING CHANGES, propriedade `delayed` removida de `Billing.php`

- adiciona especiaização de billing para pagamentos com cartão de crédito
